# Reduced redundunt copy on VM execution